fix lint error

+++ internal/fullnode/build_pods.go
@@ -18,7 +18,7 @@ func BuildPods(crd *cosmosv1.CosmosFullNode, cksums ConfigChecksums, startingOrd
 		pods    []diff.Resource[*corev1.Pod]
 	)
 	candidates := podCandidates(crd)
-	for i := int32(startingOrdinal); i < crd.Spec.Replicas; i++ {
+	for i := startingOrdinal; i < crd.Spec.Replicas; i++ {
 		pod, err := builder.WithOrdinal(i).Build()
 		if err != nil {
 			return nil, err
